+++ src/main/java/com/kh/mng/member/controller/MemberController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.kh.mng.location.model.vo.detail.Location;
+import com.kh.mng.location.model.vo.Location;
 import com.kh.mng.member.model.vo.Member;
 import com.kh.mng.member.service.MemberServiceImpl;
