+++ src/main/java/com/kh/mng/member/model/dao/MemberDao.java
@@ -3,7 +3,7 @@ package com.kh.mng.member.model.dao;
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.stereotype.Repository;
 
-import com.kh.mng.location.model.vo.detail.Location;
+import com.kh.mng.location.model.vo.Location;
 import com.kh.mng.member.model.vo.Member;
 
 @Repository
